Correções como Hash na senha e Validações nos campos, foram adicionadas e corrigidas.

+++ src/models/categoria.js
@@ -8,6 +8,9 @@ class Categoria {
 
   async inserir() {
     try {
+      if (!this.nome || this.nome.trim() === "") {
+        throw new Error("O nome da categoria é obrigatório.");
+      }
       const { db, client } = await connect();
       const result = await db.collection("categorias").insertOne({
         nome: this.nome,
@@ -15,7 +18,7 @@ class Categoria {
       console.log("Categoria inserida com sucesso:", result.insertedId);
       client.close();
     } catch (error) {
-      Logger.logError("Erro ao inserir categoria: " + error);
+      Logger.logError("Erro ao inserir categoria: " + error.message);
     }
   }
 
@@ -23,35 +26,44 @@ class Categoria {
     try {
       const { db, client } = await connect();
       const categorias = await db.collection("categorias").find(filtro).toArray();
-      console.log("Categorias encontradas com sucesso:", categorias);
+      console.log("Categoria(s) encontradas com sucesso:", categorias);
       client.close();
       return categorias;
     } catch (error) {
-      Logger.logError("Erro ao buscar categorias: " + error);
+      Logger.logError("Erro ao buscar categoria(s): " + error.message);
     }
   }
 
   static async atualizar(filtro, novosDados) {
     try {
+      if (!filtro || Object.keys(filtro).length === 0) {
+        throw new Error("Filtro obrigatório para atualização.");
+      }
+      if (!novosDados || Object.keys(novosDados).length === 0) {
+        throw new Error("Dados novos obrigatórios para atualização.");
+      }
       const { db, client } = await connect();
       const result = await db.collection("categorias").updateMany(filtro, {
         $set: novosDados,
       });
-      console.log("Categorias atualizadas com sucesso:", result.modifiedCount);
+      console.log("Categoria(s) atualizadas com sucesso:", result.modifiedCount);
       client.close();
     } catch (error) {
-      Logger.logError("Erro ao atualizar categorias: " + error);
+      Logger.logError("Erro ao atualizar categoria(s): " + error.message);
     }
   }
 
   static async deletar(filtro) {
     try {
+      if (!filtro || Object.keys(filtro).length === 0) {
+        throw new Error("Filtro obrigatório para deletar categorias.");
+      }
       const { db, client } = await connect();
       const result = await db.collection("categorias").deleteMany(filtro);
-      console.log("Categoria(a) deletadas com sucesso:", result.deletedCount);
+      console.log("Categoria(s) deletadas com sucesso:", result.deletedCount);
       client.close();
     } catch (error) {
-      Logger.logError("Erro ao deletar categorias: " + error);
+      Logger.logError("Erro ao deletar categoria(s): " + error.message);
     }
   }
 }
